+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Create a build Makefile
+- Add the `domain` and `git-url` config options
 
 ## v0.1.0 - 2019-10-10